refactor: move BuildJob schema to dedicated file and add BuildJobLogs table to database
